refactor(server): #125 — extract discovery + presence handlers into routes/{discovery,presence}.rs
