+++ src/server/mod.rs
@@ -20,18 +20,21 @@ mod ws;
 // unchanged after the #125 / WS1.4 extraction. Internal types (AppState,
 // DaemonUpdateConfig, CachedUpgradeCheck) stay private to the crate.
 use routes::{
-    add_contact, add_machine, add_task, agent_info, agent_sign, agent_user_id_handler,
-    agent_verify, announce_identity, apply_upgrade, broadcast_current_manifest, check_upgrade,
-    connect_agent, connect_diagnostics_handler, connect_machine, create_task_list,
-    delete_contact, delete_machine, direct_connections, direct_message_send_config,
-    direct_send, discovered_machine, discovered_machines, exec_cancel, exec_diagnostics,
-    exec_run, exec_sessions, file_accept_handler, file_reject_handler, file_send_handler,
-    file_transfer_send_config, file_transfer_status_handler, file_transfers_handler,
-    FileChunkAckSlot, forward_add, forward_list, forward_remove, get_a2a_agent_card,
-    get_agent_card, get_constitution, get_constitution_json, handle_file_message, health,
-    identity_revocations, identity_revoke, import_agent_card, introduction, list_contacts,
-    list_machines, list_revocations, list_task_lists, list_tasks, machines_by_user_handler,
-    pin_machine, populate_invite_base_state_from_group_info, quick_trust, revoke_contact,
+    add_contact, add_machine, add_task, agent_info, agent_reachability, agent_sign,
+    agent_user_id_handler, agent_verify, agents_by_user_handler, announce_identity,
+    apply_upgrade, broadcast_current_manifest, check_upgrade, connect_agent,
+    connect_diagnostics_handler, connect_machine, create_task_list, delete_contact,
+    delete_machine, direct_connections, direct_message_send_config, direct_send,
+    discovered_agent, discovered_agents, discovered_machine, discovered_machines, exec_cancel,
+    exec_diagnostics, exec_run, exec_sessions, file_accept_handler, file_reject_handler,
+    file_send_handler, file_transfer_send_config, file_transfer_status_handler,
+    file_transfers_handler, FileChunkAckSlot, find_agent, forward_add, forward_list,
+    forward_remove, get_a2a_agent_card, get_agent_card, get_constitution,
+    get_constitution_json, handle_file_message, health, identity_revocations, identity_revoke,
+    import_agent_card, introduction, list_contacts, list_machines, list_revocations,
+    list_task_lists, list_tasks, machine_for_agent_handler, machines_by_user_handler,
+    pin_machine, populate_invite_base_state_from_group_info, presence, presence_find,
+    presence_foaf, presence_online, presence_status, quick_trust, revoke_contact,
     run_fallback_github_poll, run_gossip_update_listener, run_startup_update_check,
     SelfPublishedReleaseManifests, shutdown_handler, status, streams_diagnostics,
     unpin_machine, update_contact, update_task, wait_for_chunk_window, wait_for_final_acks,
@@ -446,32 +449,6 @@ struct EvaluateTrustRequest {
 struct CreateWelcomeRequest {
     /// Invitee agent ID as hex string.
     agent_id: String,
-}
-
-/// Discovered identity entry from gossip announcements.
-#[derive(Debug, Serialize)]
-struct DiscoveredAgentEntry {
-    agent_id: String,
-    machine_id: String,
-    user_id: Option<String>,
-    addresses: Vec<String>,
-    announced_at: u64,
-    last_seen: u64,
-}
-
-/// Discovered machine endpoint entry from machine announcements.
-#[derive(Debug, Serialize)]
-struct DiscoveredMachineEntry {
-    machine_id: String,
-    addresses: Vec<String>,
-    announced_at: u64,
-    last_seen: u64,
-    nat_type: Option<String>,
-    can_receive_direct: Option<bool>,
-    is_relay: Option<bool>,
-    is_coordinator: Option<bool>,
-    agent_ids: Vec<String>,
-    user_ids: Vec<String>,
 }
 
 /// Peer entry.
@@ -2162,363 +2139,6 @@ async fn unsubscribe(
         not_found("subscription not found")
     }
 }
-
-/// GET /presence
-async fn presence(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    match state.agent.presence().await {
-        Ok(agents) => {
-            let entries: Vec<String> = agents.iter().map(|a| hex::encode(a.as_bytes())).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "agents": entries })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// Query parameters for presence endpoints that accept TTL and timeout.
-#[derive(Debug, Deserialize)]
-struct PresenceQueryParams {
-    /// FOAF hop count (default: 3).
-    #[serde(default = "default_foaf_ttl")]
-    ttl: u8,
-    /// Query timeout in milliseconds (default: 5000).
-    #[serde(default = "default_foaf_timeout_ms")]
-    timeout_ms: u64,
-}
-
-fn default_foaf_ttl() -> u8 {
-    3
-}
-
-fn default_foaf_timeout_ms() -> u64 {
-    5000
-}
-
-/// GET /presence/online
-///
-/// List all agents currently online (network view: all non-blocked agents from
-/// the local discovery cache).
-async fn presence_online(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    match state.agent.online_agents().await {
-        Ok(agents) => {
-            let contacts = state.agent.contacts().read().await;
-            let filtered = x0x::presence::filter_by_trust(
-                agents,
-                &contacts,
-                x0x::presence::PresenceVisibility::Network,
-            );
-            let entries: Vec<_> = filtered.into_iter().map(discovered_agent_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "agents": entries })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /presence/foaf?ttl=3&timeout_ms=5000
-///
-/// FOAF random-walk discovery of nearby agents (social view: Trusted + Known only).
-async fn presence_foaf(
-    State(state): State<Arc<AppState>>,
-    Query(params): Query<PresenceQueryParams>,
-) -> impl IntoResponse {
-    match state
-        .agent
-        .discover_agents_foaf(params.ttl, params.timeout_ms)
-        .await
-    {
-        Ok(agents) => {
-            let contacts = state.agent.contacts().read().await;
-            let filtered = x0x::presence::filter_by_trust(
-                agents,
-                &contacts,
-                x0x::presence::PresenceVisibility::Social,
-            );
-            let entries: Vec<_> = filtered.into_iter().map(discovered_agent_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "agents": entries })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /presence/find/:id?ttl=3&timeout_ms=5000
-///
-/// Find a specific agent by hex-encoded AgentId via FOAF random walk.
-async fn presence_find(
-    State(state): State<Arc<AppState>>,
-    Path(id): Path<String>,
-    Query(params): Query<PresenceQueryParams>,
-) -> impl IntoResponse {
-    let bytes = match hex::decode(&id) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("invalid agent id (expected 64 hex chars)");
-        }
-    };
-    let agent_id = x0x::identity::AgentId(bytes);
-    match state
-        .agent
-        .discover_agent_by_id(agent_id, params.ttl, params.timeout_ms)
-        .await
-    {
-        Ok(Some(agent)) => (
-            StatusCode::OK,
-            Json(serde_json::json!({ "ok": true, "agent": discovered_agent_entry(agent) })),
-        ),
-        Ok(None) => (
-            StatusCode::OK,
-            Json(serde_json::json!({ "ok": true, "agent": null })),
-        ),
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /presence/status/:id
-///
-/// Local cache lookup for a specific agent — no network I/O.
-async fn presence_status(
-    State(state): State<Arc<AppState>>,
-    Path(id): Path<String>,
-) -> impl IntoResponse {
-    let bytes = match hex::decode(&id) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("invalid agent id (expected 64 hex chars)");
-        }
-    };
-    let agent_id = x0x::identity::AgentId(bytes);
-    let cached = state.agent.cached_agent(&agent_id).await;
-    let online = cached.is_some();
-    let entry = cached.map(discovered_agent_entry);
-    (
-        StatusCode::OK,
-        Json(serde_json::json!({ "ok": true, "online": online, "agent": entry })),
-    )
-}
-
-fn discovered_agent_entry(agent: x0x::DiscoveredAgent) -> DiscoveredAgentEntry {
-    DiscoveredAgentEntry {
-        agent_id: hex::encode(agent.agent_id.as_bytes()),
-        machine_id: hex::encode(agent.machine_id.as_bytes()),
-        user_id: agent.user_id.map(|id| hex::encode(id.as_bytes())),
-        addresses: agent.addresses.into_iter().map(|a| a.to_string()).collect(),
-        announced_at: agent.announced_at,
-        last_seen: agent.last_seen,
-    }
-}
-
-fn discovered_machine_entry(machine: x0x::DiscoveredMachine) -> DiscoveredMachineEntry {
-    DiscoveredMachineEntry {
-        machine_id: hex::encode(machine.machine_id.as_bytes()),
-        addresses: machine
-            .addresses
-            .into_iter()
-            .map(|a| a.to_string())
-            .collect(),
-        announced_at: machine.announced_at,
-        last_seen: machine.last_seen,
-        nat_type: machine.nat_type,
-        can_receive_direct: machine.can_receive_direct,
-        is_relay: machine.is_relay,
-        is_coordinator: machine.is_coordinator,
-        agent_ids: machine
-            .agent_ids
-            .into_iter()
-            .map(|id| hex::encode(id.as_bytes()))
-            .collect(),
-        user_ids: machine
-            .user_ids
-            .into_iter()
-            .map(|id| hex::encode(id.as_bytes()))
-            .collect(),
-    }
-}
-
-/// GET /agents/discovered
-/// Query parameters for `GET /agents/discovered`.
-#[derive(Deserialize, Default)]
-struct DiscoveredAgentsQuery {
-    /// When `true`, return all cache entries including stale (TTL-expired).
-    #[serde(default)]
-    unfiltered: bool,
-}
-
-async fn discovered_agents(
-    State(state): State<Arc<AppState>>,
-    Query(query): Query<DiscoveredAgentsQuery>,
-) -> impl IntoResponse {
-    let result = if query.unfiltered {
-        state.agent.discovered_agents_unfiltered().await
-    } else {
-        state.agent.discovered_agents().await
-    };
-    match result {
-        Ok(agents) => {
-            let entries: Vec<DiscoveredAgentEntry> =
-                agents.into_iter().map(discovered_agent_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "agents": entries })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// Query parameters for `GET /agents/discovered/:agent_id`.
-#[derive(Deserialize, Default)]
-struct DiscoveredAgentQuery {
-    /// When `true`, wait up to 10 s for the agent to announce on its shard
-    /// topic before returning `404`. Useful for finding agents that joined
-    /// recently and may not be in cache yet.
-    #[serde(default)]
-    wait: bool,
-}
-
-/// GET /agents/discovered/:agent_id[?wait=true]
-async fn discovered_agent(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-    Query(params): Query<DiscoveredAgentQuery>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    if params.wait {
-        // Active lookup: subscribe to agent's shard, wait up to 10 s.
-        match state.agent.find_agent(agent_id).await {
-            Ok(Some(addrs)) => {
-                // Return the full discovered_agent entry if available, else
-                // synthesise a minimal response from the address list.
-                return match state.agent.discovered_agent(agent_id).await {
-                    Ok(Some(agent)) => (
-                        StatusCode::OK,
-                        Json(serde_json::json!({
-                            "ok": true,
-                            "agent": discovered_agent_entry(agent),
-                        })),
-                    ),
-                    _ => (
-                        StatusCode::OK,
-                        Json(serde_json::json!({
-                            "ok": true,
-                            "agent": {
-                                "agent_id": agent_id_hex,
-                                "addresses": addrs.iter().map(|a| a.to_string()).collect::<Vec<_>>(),
-                            }
-                        })),
-                    ),
-                };
-            }
-            Ok(None) => {
-                return not_found("agent not found within timeout");
-            }
-            Err(e) => {
-                return api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}"));
-            }
-        }
-    }
-
-    match state.agent.discovered_agent(agent_id).await {
-        Ok(Some(agent)) => (
-            StatusCode::OK,
-            Json(serde_json::json!({
-                "ok": true,
-                "agent": discovered_agent_entry(agent),
-            })),
-        ),
-        Ok(None) => not_found("agent not found"),
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /agents/:agent_id/machine
-async fn machine_for_agent_handler(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    match state.agent.machine_for_agent(agent_id).await {
-        Ok(Some(machine)) => (
-            StatusCode::OK,
-            Json(serde_json::json!({
-                "ok": true,
-                "agent_id": agent_id_hex,
-                "machine": discovered_machine_entry(machine),
-            })),
-        ),
-        Ok(None) => not_found("agent machine not found"),
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-/// GET /users/:user_id/agents
-async fn agents_by_user_handler(
-    State(state): State<Arc<AppState>>,
-    Path(user_id_hex): Path<String>,
-) -> (StatusCode, Json<serde_json::Value>) {
-    let user_id_bytes = match hex::decode(&user_id_hex) {
-        Ok(b) if b.len() == 32 => {
-            let mut arr = [0u8; 32];
-            arr.copy_from_slice(&b);
-            arr
-        }
-        _ => {
-            return bad_request("invalid user_id: expected 64 hex characters");
-        }
-    };
-    let user_id = x0x::identity::UserId(user_id_bytes);
-    match state.agent.find_agents_by_user(user_id).await {
-        Ok(agents) => {
-            let entries: Vec<DiscoveredAgentEntry> =
-                agents.into_iter().map(discovered_agent_entry).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({
-                    "ok": true,
-                    "user_id": user_id_hex,
-                    "agents": entries,
-                })),
-            )
-        }
-        Err(e) => api_error(StatusCode::INTERNAL_SERVER_ERROR, format!("{e}")),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Contact handlers
-// ---------------------------------------------------------------------------
 
 /// Parse a 64-character hex string into an AgentId.
 fn parse_agent_id_hex(hex_str: &str) -> Result<AgentId, String> {
@@ -15253,75 +14873,6 @@ async fn mls_decrypt(
 
 // ---------------------------------------------------------------------------
 // Agent discovery & connectivity handlers
-// ---------------------------------------------------------------------------
-
-/// POST /agents/find/:agent_id — actively search for an agent (3-stage: cache → shard → rendezvous).
-async fn find_agent(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    match state.agent.find_agent(agent_id).await {
-        Ok(Some(addrs)) => {
-            let addr_strs: Vec<String> = addrs.iter().map(|a| a.to_string()).collect();
-            (
-                StatusCode::OK,
-                Json(serde_json::json!({ "ok": true, "found": true, "addresses": addr_strs })),
-            )
-        }
-        Ok(None) => (
-            StatusCode::OK,
-            Json(serde_json::json!({ "ok": true, "found": false })),
-        ),
-        Err(e) => {
-            tracing::error!("find_agent failed: {e}");
-            api_error(StatusCode::INTERNAL_SERVER_ERROR, "search failed")
-        }
-    }
-}
-
-/// GET /agents/reachability/:agent_id — check reachability before connecting.
-async fn agent_reachability(
-    State(state): State<Arc<AppState>>,
-    Path(agent_id_hex): Path<String>,
-) -> impl IntoResponse {
-    let agent_id = match parse_agent_id_hex(&agent_id_hex) {
-        Ok(id) => id,
-        Err(e) => {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({ "ok": false, "error": e })),
-            );
-        }
-    };
-
-    match state.agent.reachability(&agent_id).await {
-        Some(info) => (
-            StatusCode::OK,
-            Json(serde_json::json!({
-                "ok": true,
-                "likely_direct": info.likely_direct(),
-                "needs_coordination": info.needs_coordination(),
-                "is_relay": info.is_relay(),
-                "is_coordinator": info.is_coordinator(),
-                "addresses": info.addresses.iter().map(|a| a.to_string()).collect::<Vec<_>>()
-            })),
-        ),
-        None => not_found("agent not in discovery cache"),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Contact trust extension handlers
 // ---------------------------------------------------------------------------
 
 /// POST /trust/evaluate — evaluate trust decision for an (agent, machine) pair.
