+++ src/server/routes/machines.rs
@@ -16,8 +16,10 @@ use crate as x0x;
 
 use super::super::state::AppState;
 use super::super::{
-    api_error, bad_request, discovered_machine_entry, not_found, parse_agent_id_hex,
-    parse_machine_id_hex, DiscoveredAgentQuery, DiscoveredAgentsQuery, DiscoveredMachineEntry,
+    api_error, bad_request, not_found, parse_agent_id_hex, parse_machine_id_hex,
+};
+use super::discovery::{
+    DiscoveredAgentQuery, DiscoveredAgentsQuery, DiscoveredMachineEntry, discovered_machine_entry,
 };
 use crate::contacts::MachineRecord;
 use crate::identity::MachineId;
